requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
bleach==6.3.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.1
filelock==3.20.0
Flask==2.3.2
fonttools==4.61.1
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kaggle==1.7.4.5
kiwisolver==1.4.9
lazy_loader==0.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.13.0.92
packaging @ file:///home/task_176104877067765/conda-bld/packaging_1761049113113/work
pandas==2.3.3
pillow==12.0.0
polars==1.38.1
polars-runtime-32==1.38.1
protobuf==7.34.0
psutil==7.2.2
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.25
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.3
triton==3.3.1
typer==0.24.1
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.21
ultralytics-thop==2.0.18
urllib3==2.6.3
webencodings==0.5.1
Werkzeug==2.3.6
Flask==2.3.2
Werkzeug==2.3.6
